feat: Improve the email verification process and authority management
- Added EmailVerificationScreen interface for email verification
- Implemented true system permission requests (notifications, location, camera)
- Fixed logic for checking email verification status
- Added function to resend verification email
- Optimized user registration and verification process
- Added debugging interface for problem troubleshooting
- Improved error handling and user experience

+++ backend/app/schemas/workout.py
@@ -5,7 +5,7 @@ from typing import List, Optional
 class LocationPoint(BaseModel):
     lat: float
     lng: float
-    t_ms: int  # timestamp in milliseconds
+    t_ms: float  # timestamp in milliseconds (allows sub-millisecond precision)
 
 # Start a workout
 class WorkoutStartRequest(BaseModel):
